src_code/requirements.txt
fastapi
uvicorn
torch
transformers
pydantic
numpy
Pillow
sqlalchemy
datasets
python-multipart